+++ src/Nethermind/Nethermind.Trie/BatchedTrieVisitor.cs
@@ -390,7 +390,7 @@ public readonly struct EmptyContext : INodeContext<EmptyContext>
 {
     public EmptyContext Add(ReadOnlySpan<byte> nibblePath) => this;
     public EmptyContext Add(byte nibble) => this;
-    public EmptyContext Add(TrieKey nibble) => this;
+    public EmptyContext Add(in TrieKey nibble) => this;
     public EmptyContext AddStorage(in ValueHash256 storage) => this;
 }
 
@@ -418,7 +418,7 @@ public struct TreePathContext : INodeContext<TreePathContext>
         };
     }
 
-    public TreePathContext Add(TrieKey nibble)
+    public TreePathContext Add(in TrieKey nibble)
     {
         return new TreePathContext()
         {
@@ -465,7 +465,7 @@ public readonly struct TreePathContextWithStorage : ITreePathContextWithStorage,
         };
     }
 
-    public TreePathContextWithStorage Add(TrieKey nibble)
+    public TreePathContextWithStorage Add(in TrieKey nibble)
     {
         return new TreePathContextWithStorage()
         {
@@ -500,7 +500,7 @@ public struct NoopTreePathContextWithStorage : ITreePathContextWithStorage, INod
         return this;
     }
 
-    public readonly NoopTreePathContextWithStorage Add(TrieKey nibble)
+    public readonly NoopTreePathContextWithStorage Add(in TrieKey nibble)
     {
         return this;
     }
@@ -523,6 +523,6 @@ public interface INodeContext<out TNodeContext>
 
     TNodeContext Add(byte nibble);
 
-    TNodeContext Add(TrieKey nibble);
+    TNodeContext Add(in TrieKey nibble);
     TNodeContext AddStorage(in ValueHash256 storage);
 }
